Code cleanup. Move stray declarations and functions.
