+++ vortex.vcxproj.filters
@@ -494,6 +494,9 @@
     <ClInclude Include="talents.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="player_points.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="game.def">
